Removal of ClubHome. Removal of the states and view models for the get member and remove member interface adapter.
