--- src/main/java/interface_adapter/club_home/ClubHomePresenter.java
@@ -1,4 +1,0 @@
-package interface_adapter.club_home;
-
-public class ClubHomePresenter {
-}
