+++ src/main/java/interface_adapter/club_logged_in/get_members/GetMembersPresenter.java
@@ -3,6 +3,8 @@ package interface_adapter.club_logged_in.get_members;
 import entity.data_structure.DataStore;
 import entity.user.Student;
 import interface_adapter.ViewManagerModel;
+import interface_adapter.club_logged_in.ClubLoggedInState;
+import interface_adapter.club_logged_in.ClubLoggedInViewModel;
 import use_case.club_get_members.ClubGetMembersOutputBoundary;
 import use_case.club_get_members.ClubGetMembersOutputData;
 
@@ -10,33 +12,38 @@ import use_case.club_get_members.ClubGetMembersOutputData;
  * The presenter for the get members use case for clubs.
  */
 public class GetMembersPresenter implements ClubGetMembersOutputBoundary {
-    private final GetMembersViewModel getMembersViewModel;
+    private final ClubLoggedInViewModel clubLoggedInViewModel;
     private final ViewManagerModel viewManagerModel;
 
-    public GetMembersPresenter(GetMembersViewModel getMembersViewModel,
+    public GetMembersPresenter(ClubLoggedInViewModel getMembersViewModel,
                                ViewManagerModel viewManagerModel) {
-        this.getMembersViewModel = getMembersViewModel;
+        this.clubLoggedInViewModel = getMembersViewModel;
         this.viewManagerModel = viewManagerModel;
     }
 
     @Override
     public void prepareSuccessView(ClubGetMembersOutputData data) {
         final String email = data.getEmail();
+
+        // Get the members of the club
         final DataStore<Student> members = data.getMembers();
 
-        final GetMembersState currentState = getMembersViewModel.getState();
-        currentState.setMembers(members);
-        currentState.setClubEmail(email);
-        getMembersViewModel.setState(currentState);
-        getMembersViewModel.firePropertyChanged("get members");
+        final ClubLoggedInState currentState = clubLoggedInViewModel.getState();
 
-        viewManagerModel.setState(getMembersViewModel.getViewName());
+        // Set the current members and email to the state
+        currentState.setMembers(members);
+        currentState.setEmail(email);
+        // Save the state
+        clubLoggedInViewModel.setState(currentState);
+        clubLoggedInViewModel.firePropertyChanged("get members");
+
+        viewManagerModel.setState(clubLoggedInViewModel.getViewName());
         viewManagerModel.firePropertyChanged("get members");
     }
 
     @Override
     public void prepareFailView(String errorMessage) {
-        final GetMembersState currentState = getMembersViewModel.getState();
+        final ClubLoggedInState currentState = clubLoggedInViewModel.getState();
         currentState.setErrorMessage(errorMessage);
     }
 }
