+++ src/main/java/interface_adapter/club_logged_in/remove_member/RemoveMemberPresenter.java
@@ -1,6 +1,8 @@
 package interface_adapter.club_logged_in.remove_member;
 
 import interface_adapter.ViewManagerModel;
+import interface_adapter.club_logged_in.ClubLoggedInState;
+import interface_adapter.club_logged_in.ClubLoggedInViewModel;
 import use_case.club_remove_member.ClubRemoveMemberOutputBoundary;
 import use_case.club_remove_member.ClubRemoveMemberOutputData;
 
@@ -8,31 +10,24 @@ import use_case.club_remove_member.ClubRemoveMemberOutputData;
  * The presenter for the remove member from club use case.
  */
 public class RemoveMemberPresenter implements ClubRemoveMemberOutputBoundary {
-    private final RemoveMemberViewModel removeMemberViewModel;
+    private final ClubLoggedInViewModel clubLoggedInViewModel;
     private final ViewManagerModel viewManagerModel;
 
-    public RemoveMemberPresenter(RemoveMemberViewModel removeMemberViewModel, ViewManagerModel viewManagerModel) {
-        this.removeMemberViewModel = removeMemberViewModel;
+    public RemoveMemberPresenter(ClubLoggedInViewModel removeMemberViewModel, ViewManagerModel viewManagerModel) {
+        this.clubLoggedInViewModel = removeMemberViewModel;
         this.viewManagerModel = viewManagerModel;
     }
 
     @Override
     public void prepareSuccessView(ClubRemoveMemberOutputData outputData) {
-        final RemoveMemberState removeMemberState = removeMemberViewModel.getState();
-        removeMemberState.setMemberName(outputData.getUsername());
-
-        removeMemberViewModel.setState(removeMemberState);
-        removeMemberViewModel.firePropertyChanged("remove member");
-
-        viewManagerModel.setState(removeMemberViewModel.getViewName());
-        viewManagerModel.firePropertyChanged("remove member");
+        viewManagerModel.firePropertyChanged("member removed");
     }
 
     @Override
     public void prepareFailView(String errorMessage) {
-        final RemoveMemberState removeMemberState = removeMemberViewModel.getState();
+        final ClubLoggedInState removeMemberState = clubLoggedInViewModel.getState();
         removeMemberState.setErrorMessage(errorMessage);
-        removeMemberViewModel.setState(removeMemberState);
-        removeMemberViewModel.firePropertyChanged();
+        clubLoggedInViewModel.setState(removeMemberState);
+        clubLoggedInViewModel.firePropertyChanged();
     }
 }
