--- src/main/java/interface_adapter/club_logged_in/remove_member/RemoveMemberState.java
@@ -1,45 +1,0 @@
-package interface_adapter.club_logged_in.remove_member;
-
-/**
- * The state of the data for the remove member use case for clubs.
- */
-public class RemoveMemberState {
-    private String clubEmail;
-    private String memberEmail;
-
-    private String errorMessage;
-
-    private String memberName;
-
-    public String getClubEmail() {
-        return clubEmail;
-    }
-
-    public void setClubEmail(String clubEmail) {
-        this.clubEmail = clubEmail;
-    }
-
-    public String getMemberEmail() {
-        return memberEmail;
-    }
-
-    public void setMemberEmail(String memberEmail) {
-        this.memberEmail = memberEmail;
-    }
-
-    public String getErrorMessage() {
-        return errorMessage;
-    }
-
-    public void setErrorMessage(String errorMessage) {
-        this.errorMessage = errorMessage;
-    }
-
-    public String getMemberName() {
-        return memberName;
-    }
-
-    public void setMemberName(String memberName) {
-        this.memberName = memberName;
-    }
-}
